+++ src/app/modules/alerts/store/alerts-actions.ts
@@ -43,47 +43,6 @@ export const removeAlertSuccess = createAction(
   props<{message: string}>()
 );
 
-// Alert Group Actions
-export const getAllAlertGroups = createAction (
-  '[Alert Groups] Get All',
-  props<{clientId: string}>()
-);
-
-export const getAllAlertGroupsSuccess = createAction (
-  '[Alert Groups] Get All Success',
-  props<{alertGroups: AlertGroupModel[]}>()
-);
-
-export const createAlertGroup = createAction(
-  '[Alert Groups] Create',
-  props<{clientId: string, alertGroup: AlertGroupModel}>()
-);
-
-export const createAlertGroupSuccess = createAction(
-  '[Alert Groups] Create Success',
-  props<{alertGroup: AlertGroupModel}>()
-);
-
-export const updateAlertGroup = createAction(
-  '[Alert Groups] Update',
-  props<{clientId: string, code: string, alertGroup: AlertGroupModel}>()
-);
-
-export const updateAlertGroupSuccess = createAction(
-  '[Alert Groups] Update Success',
-  props<{alertGroup: AlertGroupModel}>()
-);
-
-export const removeAlertGroup = createAction(
-  '[Alert Groups] Remove',
-  props<{clientId: string, code: string}>()
-);
-
-export const removeAlertGroupSuccess = createAction(
-  '[Alert Groups] Remove Success',
-  props<{message: string}>()
-);
-
 export const failure = createAction(
   '[Alerts] Failure',
   props<{err: {concern: 'CREATE' | 'PATCH', error: any}}>()
@@ -98,14 +57,6 @@ const all = union({
   updateAlertSuccess,
   removeAlert,
   removeAlertSuccess,
-  getAllAlertGroups,
-  getAllAlertGroupsSuccess,
-  createAlertGroup,
-  createAlertGroupSuccess,
-  updateAlertGroup,
-  updateAlertGroupSuccess,
-  removeAlertGroup,
-  removeAlertGroupSuccess,
   failure
 });
